global: back to simple file structure
* Refactors code to keep a simple file structure (closes #62)
