+++ .gitignore
@@ -20,3 +20,5 @@ __pycache__/
 .tox
 build
 _build
+
+claimstore/static/vendors
