+++ claimstore/static/json/examples/claim.cds.json
