+++ claimstore/static/json/examples/claimant.cds.json
