+++ claimstore/static/json/schemas/predicate.json
