+++ claimstore/templates/contact.html
@@ -18,27 +18,13 @@
 # 59 Temple Place, Suite 330, Boston, MA 02111-1307, USA.
 #}
 
-{% extends "claims/base.html" %}
+{% extends "base.html" %}
 
 {% block content %}
 
 <div>
-    <h1>
-        Oops!</h1>
-    <h2>
-        404 Not Found</h2>
-    <div class="error-details">
-        Requested page not found!
-    </div>
-    <div class="error-actions">
-        <a href="{{ url_for('claims_views.index') }}">
-          <span class="glyphicon glyphicon-home"></span>
-          Take Me Home
-        </a>
-        <a href="{{ url_for('claims_views.contact') }}">
-          <span class="glyphicon glyphicon-envelope"></span> Contact Support
-        </a>
-    </div>
+  <h1>Contact</h1>
+  <p class="lead">To be done</p>
 </div>
 
 {% endblock %}
