+++ docs/api/claimstore.models.rst
@@ -1,0 +1,7 @@
+claimstore.models module
+========================
+
+.. automodule:: claimstore.models
+    :members:
+    :undoc-members:
+    :show-inheritance:
