--- docs/api/claimstore.modules.claims.cli.rst
@@ -1,7 +1,0 @@
-claimstore.modules.claims.cli module
-====================================
-
-.. automodule:: claimstore.modules.claims.cli
-    :members:
-    :undoc-members:
-    :show-inheritance:
