--- docs/api/claimstore.modules.claims.config.rst
@@ -1,7 +1,0 @@
-claimstore.modules.claims.config module
-=======================================
-
-.. automodule:: claimstore.modules.claims.config
-    :members:
-    :undoc-members:
-    :show-inheritance:
