--- docs/api/claimstore.modules.claims.models.rst
@@ -1,7 +1,0 @@
-claimstore.modules.claims.models module
-=======================================
-
-.. automodule:: claimstore.modules.claims.models
-    :members:
-    :undoc-members:
-    :show-inheritance:
