--- docs/api/claimstore.modules.claims.restful.rst
@@ -1,7 +1,0 @@
-claimstore.modules.claims.restful module
-========================================
-
-.. automodule:: claimstore.modules.claims.restful
-    :members:
-    :undoc-members:
-    :show-inheritance:
