--- docs/api/claimstore.modules.claims.rst
@@ -1,21 +1,0 @@
-claimstore.modules.claims package
-=================================
-
-Submodules
-----------
-
-.. toctree::
-
-   claimstore.modules.claims.cli
-   claimstore.modules.claims.config
-   claimstore.modules.claims.models
-   claimstore.modules.claims.restful
-   claimstore.modules.claims.views
-
-Module contents
----------------
-
-.. automodule:: claimstore.modules.claims
-    :members:
-    :undoc-members:
-    :show-inheritance:
