--- docs/api/claimstore.modules.claims.views.rst
@@ -1,7 +1,0 @@
-claimstore.modules.claims.views module
-======================================
-
-.. automodule:: claimstore.modules.claims.views
-    :members:
-    :undoc-members:
-    :show-inheritance:
