--- docs/api/claimstore.modules.rst
@@ -1,17 +1,0 @@
-claimstore.modules package
-==========================
-
-Subpackages
------------
-
-.. toctree::
-
-    claimstore.modules.claims
-
-Module contents
----------------
-
-.. automodule:: claimstore.modules
-    :members:
-    :undoc-members:
-    :show-inheritance:
