+++ docs/api/claimstore.restful.rst
@@ -1,0 +1,7 @@
+claimstore.restful module
+=========================
+
+.. automodule:: claimstore.restful
+    :members:
+    :undoc-members:
+    :show-inheritance:
