+++ docs/api/claimstore.views.rst
@@ -1,0 +1,7 @@
+claimstore.views module
+=======================
+
+.. automodule:: claimstore.views
+    :members:
+    :undoc-members:
+    :show-inheritance:
